+++ packages/sdk-components-react/src/components.ts
@@ -30,4 +30,3 @@ export { Vimeo } from "./vimeo";
 export { VimeoPreviewImage } from "./vimeo-preview-image";
 export { VimeoPlayButton } from "./vimeo-play-button";
 export { VimeoSpinner } from "./vimeo-spinner";
-export { MarketplaceItem } from "./marketplace-item";
