+++ packages/sdk-components-react/src/metas.ts
@@ -30,4 +30,3 @@ export { meta as Vimeo } from "./vimeo.ws";
 export { meta as VimeoPreviewImage } from "./vimeo-preview-image.ws";
 export { meta as VimeoPlayButton } from "./vimeo-play-button.ws";
 export { meta as VimeoSpinner } from "./vimeo-spinner.ws";
-export { meta as MarketplaceItem } from "./marketplace-item.ws";
